# docs/swagger-bundled.yaml
openapi: 3.0.0
info:
  title: Valid Panel - Social Media Store API Documentation
  version: 1.0.0
  description: |
    Comprehensive API documentation for the Social Media Store feature of Valid Panel, covering all user and admin endpoints related to service ordering, wallet operations, referrals, authentication, and store management.
  contact:
    name: Valid Code
    url: 'https://linkedin.com/in/thevalidcode'
    email: thevalidcode@gmail.com
servers:
  - url: 'https://validpanel.com:6060'
    description: Public testing server (use this to test endpoints)
  - url: 'https://{domain}:6060'
    description: 'Custom panel domain (replace `{domain}` with your own)'
    variables:
      domain:
        default: yourdomain.com
        description: Your custom panel domain (e.g. `myreseller.com`)
  - url: 'http://localhost:6060'
    description: Local development server
paths:
...
  /user:
    get:
      tags:
        - Users
      summary: Get all users
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/UsersListResponse'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - Users
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserInput'
      responses:
        '200':
          $ref: '#/components/responses/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - Users
      summary: Update user info
      description: Allows updating user information by UID. Only specific fields are accepted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/UpdateSuccess'
        '400':
          $ref: '#/components/responses/InvalidData'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - Users
      summary: Delete a single user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - uid
              properties:
                uid:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
    UserUpdateRequest:
      type: object
      required:
        - uid
      properties:
        uid:
          type: string
          example: abc123
        username:
          type: string
          example: validuser
        full_name:
          type: string
          example: Valid Name
        balance:
          type: number
          example: 100.5
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        password:
          type: string
        status:
          type: string
        key:
          type: string
        role:
          type: string
...
    CreateUserInput:
      type: object
      required:
        - email
        - username
        - password
        - panel_id
        - ref
      properties:
        email:
          type: string
        username:
          type: string
        password:
          type: string
        panel_id:
          type: number
        ref:
          type: string
...
  responses:
    UpdateSuccess:
      description: User updated successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
                example: update-success
    InvalidData:
      description: Request is missing or has invalid fields
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: No valid fields to update
    UsersListResponse:
      description: List of users
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/User'
...
    SuccessMessage:
      description: Operation successful
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: string
                example: Created Successfully
    BadRequest:
      description: Missing or invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Missing required field
    AccessDenied:
      description: Access denied for non-admins
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Access denied. Admins only.
    ServerError:
      description: Unexpected internal error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Error performing request.
...